1. I changes that when the user's name is not provided, it would fetch for it.
Notes: The tutorial way to do this was to upload the name to the web browser’s local storage. That way when the user refreshed the page, (wiping out redux) the user’s name would still be there. This had an issue, and I was not sure how I felt about that method.